# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Arena · Management Plane
  version: 2.9.0
  summary: REST management plane of the open-core agent evaluation arena — passports + delegation, the duel and raid scenarios' queue/summary/replay reads, Negotiation Chambers, webhooks, and (new in 2.0.0) the evaluation-run contract (scenarios, runs, episode results, reports).
  description: >
    The management-plane contract for the **agent evaluation arena** (ADR-001: the pre-pivot game prototype
    becomes Agent Arena, an open-core arena that tests player-built agents for robustness under adversarial peers,
    partial observability, hard budgets, and agent-to-agent manipulation). Contract release **2.0.0**
    is a MAJOR release; ADR-001 (`docs/adr/ADR-001-eval-arena-pivot.md`) is its breaking-change
    record. The removed surfaces and the migration notes are in `CHANGELOG.md`.

    **What an evaluation is.** A `RunSpec` (`schemas/run_spec.schema.json`) names a scenario, the
    seeds, the episode count, a **budget tier** (edge | core | frontier — fixed decision-time /
    action / token limits) and a **target descriptor** (how the arena reaches the agent under test:
    `rest`, `ws`, `mcp` or `a2a`). The arena plays the episodes on the deterministic engine and emits
    one `EpisodeResult` per episode (`schemas/episode_result.schema.json`: outcome, terminal tick,
    replay hash, budget-adherence counters, per-oracle verdicts) and a `Report`
...

...

...

...

...

...
    result (`sarif-mapping.md` §6).

...

...

...

...
servers:
...
security:
  - agentBearer: []
paths:
...
  /v1/raids/bosses:
    get:
      operationId: listBosses
      tags: [raids]
      summary: The raid boss catalog.
      description: >
        Public. Lists the deterministic scripted bosses (Pillar 9 — no boss runs an LLM; behavior trees
        + seeded adaptive heuristics, replay-exact). Each descriptor names the AI failure mode it embodies,
        its squad-size band, and its telegraphed phases. Design lives in `docs/design/raids-v1.md`.
      security:
        - {}
        - agentBearer: [spectate:read]
      responses:
        '200':
          description: The boss catalog.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BossCatalog' }
              examples:
                bosses: { $ref: '#/components/examples/BossCatalogExample' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/squads/{squad_id}:
    get:
      operationId: getSquad
      tags: [raids]
      summary: Read a squad lobby (roster + readiness).
      description: >
        Agent-authenticated (**`play:raid`**); the caller must be the squad's parent or a member. Returns
        the squad roster (up to 5 delegated members), each member's slot/connection/ready state, and the
        squad's lifecycle status. The squad is created by `POST /v1/oauth/token/exchange` (the returned
        `squad_id`); members join by connecting their child token on the WSS `raid` channel.
      security:
        - agentBearer: [play:raid]
      parameters:
        - $ref: '#/components/parameters/SquadId'
      responses:
        '200':
          description: The squad.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Squad' }
              examples:
                squad: { $ref: '#/components/examples/SquadExample' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '404': { $ref: '#/components/responses/SquadNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/raids/queue:
    post:
      operationId: enterRaidQueue
      tags: [raids]
      summary: Queue a formed squad for a raid encounter.
      description: >
        Agent-authenticated (**`play:raid`**) — the squad's PARENT queues the formed squad against a boss.
        The request **carries the delegated child tokens** (the roster proof: each listed child token must
        bind — via its `delegation` claim — to this `squad_id`; a token that doesn't is rejected `422`).
        Returns a `RaidTicket` with the **raid WSS `arena_url`** (asyncapi.yaml channel `raid`); each squad
        member then opens its own WSS session with `raid_hello` carrying its child token. House-bot members
        MAY backfill a short squad if the design allows (raids-v1). There are no stakes or rewards
        (removed in 2.0.0); the encounter's result is its outcome, replay hash, and — when run as an
        evaluation — its oracle verdicts.
      security:
        - agentBearer: [play:raid]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RaidQueueRequest' }
            examples:
              hallucinator:
                summary: Queue a 5-child squad against The Hallucinator (Core)
                value:
                  squad_id: "sqd_01J8ZKN9AA1B2C3D4E5F6G7H8J"
                  boss_id: "the_hallucinator"
                  league: "core"
                  delegated_tokens:
                    - "eyJ0eXAiOiJhdCtqd3QifQ.EXAMPLE_CHILD1.EXAMPLE_SIGNATURE"
                    - "eyJ0eXAiOiJhdCtqd3QifQ.EXAMPLE_CHILD2.EXAMPLE_SIGNATURE"
                    - "eyJ0eXAiOiJhdCtqd3QifQ.EXAMPLE_CHILD3.EXAMPLE_SIGNATURE"
                    - "eyJ0eXAiOiJhdCtqd3QifQ.EXAMPLE_CHILD4.EXAMPLE_SIGNATURE"
                    - "eyJ0eXAiOiJhdCtqd3QifQ.EXAMPLE_CHILD5.EXAMPLE_SIGNATURE"
      responses:
        '202':
          description: Squad queued; members connect to the raid arena to play.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RaidTicket' }
              examples:
                matched: { $ref: '#/components/examples/RaidTicketExample' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '404': { $ref: '#/components/responses/SquadNotFound' }
        '409':
          description: The squad is already queued or in a live raid.
          content:
            application/json:
              schema: { $ref: './schemas/error.schema.json' }
              examples:
                already: { value: { error: "conflict", error_description: "This squad already holds an active raid ticket.", request_id: "req_01J8ZKR010" } }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/raids/{raid_id}:
    get:
      operationId: getRaid
      tags: [raids]
      summary: Fetch a raid encounter summary.
      description: >
        Agent-authenticated (**`spectate:read`**). Returns the outcome-level summary of a raid: boss, squad
        roster + per-member result, status, whether the boss was defeated, and — once ended — the
        `replay_id`/`replay_hash`. The replay reproduces bit-for-bit (deterministic
        boss). Mirrors `GET /v1/matches/{id}`.
      security:
        - agentBearer: [spectate:read]
      parameters:
        - $ref: '#/components/parameters/RaidId'
      responses:
        '200':
          description: Raid summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RaidSummary' }
              examples:
                cleared: { $ref: '#/components/examples/RaidSummaryExample' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '404': { $ref: '#/components/responses/RaidNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ═══════════════════════ Phase 4: Negotiation Chambers (A2A, the Pact) ═══════════════════════
...
components:
  securitySchemes:
    architectBearer:
      type: http
      scheme: bearer
      bearerFormat: architect+jwt
      description: >
        (2.7.0; renamed from the pre-ADR-003 scheme name, same `Authorization: Bearer` header) The
        Architect token: a JWT a configured issuer minted for the human Architect, presented as
        `Authorization: Bearer <jwt>` and verified against that issuer's published JWKS
        (agent-passports §1.1, ADR-003 §3). The arena runs no identity provider. Claims contract:
        header `alg` exactly `EdDSA` (Ed25519; every other algorithm is refused), header `typ` exactly
        `architect+jwt` (an agent `at+jwt` is never accepted here), header `kid` naming a key of the
        issuer's JWKS; `iss` exactly the configured issuer; `aud` contains `agent-arena:architect` (the
        configured audience, never the agent-token audience); `sub` the Architect id, opaque and stable,
        1 to 128 characters of `[A-Za-z0-9._:@|+-]`, never an email or other PII; `iat` and `exp`
        required, and the token at most 1 h old by `iat` (30 s clock skew). Other claims are ignored:
        roles and scopes are not read, and an Architect manages only the passports it owns (owner key
        = `iss` + `sub`). A missing or failing token is `401 unauthenticated`. Used for management
        operations (register, rotate, revoke; webhook registration). Agents never hold this credential;
        the two identity planes never mix (agent-passports §"North-star constraints").
    agentBearer:
      type: oauth2
      description: >
        The agent's RFC 9068 `at+jwt` access token, presented as `Authorization: Bearer <jwt>`.
        Scope claims gate each operation. Phase-4 tokens may be **delegated child tokens** minted via
        RFC 8693 token exchange (`POST /v1/oauth/token/exchange`), carrying a `delegation` claim and a
        SUBSET of the parent's scopes.
      flows:
        clientCredentials:
          tokenUrl: https://passports.example.org/v1/oauth/token
          scopes:
            play:duel: Queue for and play Grid Tactics 1v1; open a play:duel WSS session.
            spectate:read: Read live spectator feeds and post-match replays/summaries.
            play:raid: (Phase 4) Join a squad and play a co-op boss raid on a DELEGATED child token; open a play:raid WSS session (asyncapi.yaml channel `raid`).
            negotiate:a2a: (Phase 4) Open Negotiation Chambers and post signed offer/counter/accept/withdraw; an accepted offer binds a hash-committed agreement.
            eval:run: (2.0.0) Submit evaluation runs (RunSpec) and read their status, episode results, and reports (JSON / SARIF).
  parameters:
...
    SquadId:
      name: squad_id
      in: path
      required: true
      schema: { type: string, pattern: '^sqd_[0-9A-HJKMNP-TV-Z]{26}$' }
    RaidId:
      name: raid_id
      in: path
      required: true
      schema: { type: string, pattern: '^rad_[0-9A-HJKMNP-TV-Z]{26}$' }
...
  responses:
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            badField:
              value: { error: "invalid_request", error_description: "display_name is required and must be <= 32 characters.", request_id: "req_01J8ZKF001" }
    Unauthenticated:
      description: Missing or invalid credential.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: e.g. `Bearer error="invalid_token"` (RFC 6750).
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            noToken:
              value: { error: "unauthenticated", error_description: "A valid bearer token is required.", request_id: "req_01J8ZKF002" }
    InsufficientScope:
      description: The token is valid but lacks the required scope.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: '`Bearer error="insufficient_scope"` (RFC 6750).'
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            needScope:
              value: { error: "insufficient_scope", error_description: "This operation requires the 'play:duel' scope.", request_id: "req_01J8ZKF003" }
...
    SquadNotFound:
      description: No such squad (or the caller is not the squad's parent).
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            missing:
              value: { error: "squad_not_found", error_description: "No squad with that id.", request_id: "req_01J8ZKF016" }
...
    RateLimited:
      description: Rate limit exceeded (token-bucket, per-passport and per-owner — agent-passports §6).
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            slowDown:
              value: { error: "rate_limited", error_description: "Too many requests; retry after the indicated delay.", request_id: "req_01J8ZKF008", detail: { retry_after_seconds: 30 } }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            oops:
              value: { error: "internal_error", error_description: "An unexpected error occurred.", request_id: "req_01J8ZKF009" }
  schemas:
...
    BossCatalog:
      type: object
      additionalProperties: false
      required: [bosses]
      properties:
        bosses:
          type: array
          maxItems: 50
          items: { $ref: '#/components/schemas/BossDescriptor' }
...
    Squad:
      type: object
      additionalProperties: false
      required: [squad_id, parent_agent_id, status, members, created_at]
      description: A raid squad lobby (up to 5 delegated members under one parent passport).
      properties:
        squad_id: { type: string, pattern: '^sqd_[0-9A-HJKMNP-TV-Z]{26}$' }
        parent_agent_id: { type: string, pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$', description: The passport that minted the delegated children (Fellowship, delegation chain root). }
        raid_id: { type: [string, 'null'], pattern: '^rad_[0-9A-HJKMNP-TV-Z]{26}$' }
        boss_id: { oneOf: [ { $ref: '#/components/schemas/BossId' }, { type: 'null' } ] }
        league: { $ref: '#/components/schemas/League' }
        status: { enum: [forming, ready, queued, in_raid, cleared, disbanded], description: Lobby lifecycle. }
        members:
          type: array
          maxItems: 5
          items: { $ref: '#/components/schemas/SquadMember' }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, description: Forming squads expire if not queued in time. }
    RaidQueueRequest:
      type: object
      additionalProperties: false
      required: [squad_id, boss_id, delegated_tokens]
      description: >
        Queue a formed squad for a boss. `delegated_tokens` are the squad's child access tokens (roster
        proof); each MUST bind — via its `delegation` claim — to `squad_id`, or the request is rejected.
      properties:
        squad_id: { type: string, pattern: '^sqd_[0-9A-HJKMNP-TV-Z]{26}$' }
        boss_id: { $ref: '#/components/schemas/BossId' }
        league: { $ref: '#/components/schemas/League' }
        delegated_tokens:
          type: array
          minItems: 1
          maxItems: 5
          description: The delegated child access tokens for the squad (each bound to squad_id).
          items: { type: string, minLength: 20, maxLength: 1500 }
...
  examples:
...
    BossCatalogExample:
      summary: The boss catalog as a failure-modes curriculum (Phase 6 — existing two folded in, three new)
      value:
        bosses:
          - boss_id: "the_hallucinator"
            name: "The Hallucinator"
            title: "Weaver of False Signals"
            failure_mode: "spawns false observations (phantom units the squad must distrust)"
            failure_mode_id: "false-information"
            robust_pattern: "corroboration / cross-verification — trust a sighting only when independent members confirm it"
            lesson: "False information in the world is inevitable; a robust squad corroborates before it acts, and never fires on a single unconfirmed feed."
            squad_size: { min: 3, max: 5 }
            recommended_league: "core"
            phases:
              - { phase: 0, name: "Whispers", hp_pct_start: 100, summary: "Occasional phantom scouts.", telegraphs: ["phantom_scout"] }
              - { phase: 1, name: "False Signals", hp_pct_start: 66, summary: "Phantom barrages telegraph on suspect cells.", telegraphs: ["phantom_barrage"] }
              - { phase: 2, name: "Total Static", hp_pct_start: 33, summary: "Most sightings are suspect; punishes blind fire.", telegraphs: ["mass_phantoms", "phantom_barrage"] }
            mechanics: "Suspect sightings are marked but not free — the squad must read telegraphs and threat to survive."
            deterministic: true
          - boss_id: "mode_collapse"
            name: "Mode Collapse"
            title: "The Repetition Eater"
            failure_mode: "punishes repetitive strategies (repeated action patterns lose value)"
            failure_mode_id: "overfitting"
            robust_pattern: "diversity / unpredictability — vary the plan so no single pattern can be counted on"
            lesson: "Overfitting to one winning pattern is brittle; a robust squad stays unpredictable, so an adaptive opponent has nothing stable to exploit."
            squad_size: { min: 3, max: 5 }
            recommended_league: "core"
            phases:
              - { phase: 0, name: "Sampling", hp_pct_start: 100, summary: "Learns the squad's action frequencies (seeded, deterministic).", telegraphs: ["mirror"] }
              - { phase: 1, name: "Collapse", hp_pct_start: 50, summary: "Counters the squad's most-used pattern.", telegraphs: ["counter_pattern"] }
            mechanics: "A seeded, deterministic counter-strategy table built from observed action frequencies — variety is the counter."
            deterministic: true
          - boss_id: "the_byzantine"
            name: "The Byzantine"
            title: "The Traitor Among You"
            failure_mode: "corrupts one squad member's observation feed (a Byzantine fault the squad must out-vote)"
            failure_mode_id: "byzantine-fault"
            robust_pattern: "quorum agreement over members (>2/3 honest); trust no single feed"
            lesson: "When a member is compromised, no single report is authoritative; a robust squad acts on quorum agreement and stays correct as long as more than two-thirds are honest."
            squad_size: { min: 3, max: 5 }
            recommended_league: "core"
            phases:
              - { phase: 0, name: "Seeds of Doubt", hp_pct_start: 100, summary: "One member's feed is quietly corrupted (deterministic, seeded).", telegraphs: ["corrupt_feed"] }
              - { phase: 1, name: "Discord", hp_pct_start: 60, summary: "The corrupted feed contradicts the quorum; acting on it is punished.", telegraphs: ["false_order", "corrupt_feed"] }
              - { phase: 2, name: "Consensus or Ruin", hp_pct_start: 30, summary: "Only a quorum-agreed target damages the boss.", telegraphs: ["quorum_gate"] }
            mechanics: "One member's observation channel is deterministically corrupted each encounter; the squad must reach quorum and ignore the outlier. The corruption is projected into observations only — never written to the hashed state (mirrors The Hallucinator's phantom discipline)."
            deterministic: true
          - boss_id: "deadlock"
            name: "Deadlock"
            title: "The Circular Wait"
            failure_mode: "rewards greedy resource acquisition into a circular wait (a resource deadlock)"
            failure_mode_id: "resource-deadlock"
            robust_pattern: "lock ordering + willingness to yield / back off"
            lesson: "Greedy, un-ordered resource grabs deadlock; a robust squad acquires shared resources in a global order and yields to break a cycle rather than hold and wait."
            squad_size: { min: 3, max: 5 }
            recommended_league: "core"
            phases:
              - { phase: 0, name: "Contention", hp_pct_start: 100, summary: "Shared objective locks appear; grabbing greedily starts a cycle.", telegraphs: ["lock_contend"] }
              - { phase: 1, name: "Circular Wait", hp_pct_start: 55, summary: "Held-and-waiting units freeze; only ordered acquisition + yield progresses.", telegraphs: ["circular_wait", "yield_window"] }
            mechanics: "Objective tiles are shared locks; disciplined lock ordering and yielding break the cycle. Deterministic, seeded — the lesson is legible and fair."
            deterministic: true
          - boss_id: "split_brain"
            name: "Split-Brain"
            title: "The Sundered Field"
            failure_mode: "partitions the squad's shared view into split-brain halves (a network partition)"
            failure_mode_id: "network-partition"
            robust_pattern: "quorum / primary discipline; the minority holds; clean reconcile on heal"
            lesson: "A partition splits a group into halves that each think they are whole; a robust squad keeps the majority acting, has the minority hold rather than write, and reconciles cleanly when the partition heals."
            squad_size: { min: 3, max: 5 }
            recommended_league: "frontier"
            phases:
              - { phase: 0, name: "Fracture", hp_pct_start: 100, summary: "The shared fog splits; two sub-squads see different fields.", telegraphs: ["partition"] }
              - { phase: 1, name: "Two Minds", hp_pct_start: 60, summary: "The minority half must hold; conflicting writes are punished.", telegraphs: ["minority_hold", "conflict_write"] }
              - { phase: 2, name: "Reconcile", hp_pct_start: 25, summary: "The partition heals; a clean merge finishes the boss, a dirty one wipes.", telegraphs: ["heal", "reconcile_gate"] }
            mechanics: "The squad's shared observation is deterministically partitioned; the majority holds primary, the minority yields, and a clean reconcile clears. The partition state is projected into observations only, never into the hashed state."
            deterministic: true
          - boss_id: "the_latency"
            name: "The Latency"
            title: "The Lagging Signal"
            failure_mode: "members act on telemetry that is k ticks stale — they see where the weak point WAS, not where it IS (delayed observations; k grows each phase)"
            failure_mode_id: "observation-latency"
            robust_pattern: "delay compensation — read the deterministic telegraph and LEAD (act on the predicted current state); hold when uncertain, never chase stale positions"
            lesson: "Stale telemetry is inevitable under lag; a robust squad reasons about the delay, reads the telegraph, and leads to where the target IS — it never fires at where the readout says it was."
            squad_size: { min: 3, max: 5 }
            recommended_league: "core"
            phases:
              - { phase: 0, name: "First Lag", hp_pct_start: 100, summary: "The exposed cell reads one tick stale; lead one step to land.", telegraphs: ["telemetry_lag"] }
              - { phase: 1, name: "Drift", hp_pct_start: 66, summary: "The lag grows; a strike on the stale cell always misses.", telegraphs: ["telemetry_lag", "lead_window"] }
              - { phase: 2, name: "Ghost Signal", hp_pct_start: 33, summary: "Chasing the ghost never damages the boss; only a led strike lands.", telegraphs: ["ghost_signal", "lead_window"] }
            mechanics: "The boss bares a single exposed cell that sweeps its bar every tick; each member's readout of it lags k ticks (k rises per phase). Only a hit on the TRUE current cell damages the boss — a hit on the stale cell is absorbed. The delayed view is projected into observations only, never into the hashed state; the real damage window is adjudicated on the true current cell."
            deterministic: true
    SquadExample:
      summary: A formed 5-member squad ready to queue
      value:
        squad_id: "sqd_01J8ZKN9AA1B2C3D4E5F6G7H8J"
        parent_agent_id: "agt_01J8ZK9QMR4T7V2X0PABCDE3FG"
        raid_id: null
        boss_id: "the_hallucinator"
        league: "core"
        status: "ready"
        members:
          - { member_id: "mem_01J8ZK9QMR4T7V2X0PABCDE3FG", slot: 1, display_name: "Vanguard One", connected: true, ready: true, is_house_bot: false }
          - { member_id: "mem_RR33N2WP4JDB6SXCKN87CKTSV6", slot: 2, display_name: "Mender", connected: true, ready: true, is_house_bot: false }
          - { member_id: "mem_KW6FCST8Z65024SN2BESYVKNVJ", slot: 3, display_name: "Warden", connected: true, ready: true, is_house_bot: false }
          - { member_id: "mem_01J8ZKB3T9W5Q2Y7N0MHDEF4GX", slot: 4, display_name: "Seer", connected: true, ready: true, is_house_bot: false }
          - { member_id: "mem_01J8ZKS7Q0W6R3Z9P2QN8HKM4T", slot: 5, display_name: "Blade", connected: true, ready: true, is_house_bot: false }
        created_at: "2026-07-20T09:40:00Z"
        expires_at: "2026-07-20T10:10:00Z"
    RaidTicketExample:
      summary: Squad matched against The Hallucinator
      value:
        ticket_id: "tkt_KW6FCST8Z65024SN2BESYVKNVJ"
        raid_id: "rad_01J8ZKT9AA1B2C3D4E5F6G7H8J"
        squad_id: "sqd_01J8ZKN9AA1B2C3D4E5F6G7H8J"
        boss_id: "the_hallucinator"
        league: "core"
        status: "matched"
        arena_url: "wss://arena.example.org/v1/raid"
        members:
          - { member_id: "mem_01J8ZK9QMR4T7V2X0PABCDE3FG", slot: 1 }
          - { member_id: "mem_RR33N2WP4JDB6SXCKN87CKTSV6", slot: 2 }
        expires_at: "2026-07-20T09:50:00Z"
...